Revert "Added scheduling and memory management as well as saving before attempting to switch routines."
This reverts commit 0995e6b20877c6f0a7d473a4ab3fed020c1a48b9.
