+++ config.json
@@ -1,10 +1,10 @@
 {
-  "routine": 2,
+  "routine": 1,
   "mode": 1,
-  "name": "ILLO",
+  "name": "ILLO_1",
   "college": "penn_state",
   "college_spirit_enabled": true,
   "college_chant_detection_enabled": false,
-  "ufo_persistent_memory": true,
-  "bluetooth_enabled": true
+  "ufo_persistent_memory": false,
+  "bluetooth_enabled": false
 }
